Refactor and cleanup, no functional change.
- Move a bunch of stuff around to increase clarity.

- Delete test flake-related scafolding, since it is being
replaced with something different.
