+++ policybot/handlers/githubwebhook/filters/cfgmonitor/monitor.go
@@ -20,12 +20,10 @@ import (
 	"net/http"
 	"strings"
 
-	"istio.io/bots/policybot/pkg/gh"
-
-	"istio.io/bots/policybot/pkg/fw"
-
 	webhook "github.com/go-playground/webhooks/github"
 
+	"istio.io/bots/policybot/handlers/githubwebhook/filters"
+	"istio.io/bots/policybot/pkg/gh"
 	"istio.io/pkg/log"
 )
 
@@ -42,7 +40,7 @@ type Monitor struct {
 
 var scope = log.RegisterScope("monitor", "Listens for changes in policybot config", 0)
 
-func NewMonitor(ctx context.Context, ght *gh.ThrottledClient, repo string, file string, notify func()) (fw.Webhook, error) {
+func NewMonitor(ctx context.Context, ght *gh.ThrottledClient, repo string, file string, notify func()) (filters.Filter, error) {
 	if repo == "" {
 		// disable everything if we don't have a repo
 		return &Monitor{}, nil
